add downloading (#382)
